This commit is for Chapter 3 and its listings. I also edited the index page to describe what I have going on.
I have a route for /
This route will take the path and try to render it.
EX:
  localhost:3000/listing3-10

Will take everything after the / (aka the path)
And attempt to render a file with that name.
EX: listing3-10.ejs

This makes it so I can create new HTML files on the fly as I read a chapter, create
a CSS file to go with it, and be able to access it via the name of the file right away instead
of having to create a new route for it.

Not sure if this will work the whole way for the book, like maybe the listings will get more
complex. We shall see.